prospect-pipeline/requirements.txt
httpx>=0.27
pandas>=2.1
jinja2>=3.1
openpyxl>=3.1
beautifulsoup4>=4.12
pyyaml>=6.0
click>=8.1
anthropic>=0.40
python-dotenv>=1.0
tenacity>=8.2
rapidfuzz>=3.6
# feedparser is optional — used if you extend press ingestion to RSS. Pin
# carefully on Python 3.12+ (sgmllib3k dep can fail to build).

